new_dynarec: ARM64 fixed set_jump_target and emit_read_ptr Implemented following emitters: emit_sbcs emit_and emit_sltiu32 emit_cmp emit_set_if_less64_32 emit_set_if_carry64_32 inline_readstub inline_writestub do_miniht_insert

+++ src/r4300/new_dynarec/arm64/arm64_profiler.h
@@ -102,7 +102,7 @@ typedef struct
   uint64_t PC; //precomp_instr * PC;
   precomp_instr fake_pc;
   uint64_t ram_offset;
-  u_int mini_ht[32][2];
+  uint64_t mini_ht[32][2];
   u_char restore_candidate[512];
   uint64_t memory_map[1048576];
 }profiler_t;
